Also add version param in debug mode
Modifies `_make_output_url` to accept a url to a specific file so we can calculate the version even when debug mode is on. This is helpful because it means browser caching will be bypassed when debug mode is on. As my static `css` and `js` files are being developed, I can now see their full source in developer console, and always be sure when I reload that the browser is displaying the newest version.